Added support for blocks when defining a class
Since `SimpleOperation.new` always returns a new class, there is no need
to inherit from that class, we can simply assign it to a constant.

Now it's possible to replace `class CreateUser < SimpleOperation.new`
with `CreateUser = SimpleOperation.new`. This syntax removes the anonymous
class from list of ancestors (it was confusing).

The only thing that changes in this syntax is that we need to add `do` after
`SimpleOperation.new`, since we're not using a Ruby class syntax anymore,
but passing a block to a method.